An implementation for B2G to ask allow users to log in even if Persona has not verified that the user owns the email address.
Starts with navigator.id.request({ allowUnverified: true });
When the assertion is returned to the RP, and if they are using
the Persona verifier, must be sure to include an allowUnverified=true
parameter to the verifier, otherwise it will be rejected.

Behind the scenes, when the dialog tries to stage the user, instead
the user is created immediately, an email is added in our email table,
and a stagedEmail token is created and send to the user. When
allowUnverified is true, our wsapi will also check for a staged email/pass,
and use that to allow the user to login.

(merged from commit ebe21f3e3fc9bb80f65f694b359e377ddfad64a0)

Conflicts:
	example/rp/index.html